requirements.txt
pandas
playwright
google-cloud-storage
google-cloud-bigquery
